+++ pyproject.toml
@@ -10,6 +10,7 @@ python = "^3.12"
 protobuf = "^4.25.1"
 websockets = "^12.0"
 numpy = "^1.26.2"
+pillow = "^10.1.0"
 
 [build-system]
 requires = ["poetry-core"]
