+++ tests/test1.py
@@ -10,7 +10,7 @@ import numpy as np
 
 
 async def main():
-    tree = libacmchristmas.TreeController(url=os.environ["TEST_URL"])
+    tree = libacmchristmas.tree.TreeController(url=os.environ["TEST_URL"])
     await tree.connect()
     for n in range(20):
         tree.leds[n] = 0x888800 + 8*n
@@ -21,8 +21,8 @@ async def main():
     img = np.array(object=[], dtype=np.uint8)
     for x in range(tree.ix):
         for y in range(tree.iy):
-            c = math.floor(((x+y)/(tree.ix+tree.iy))*256)
-            img = np.concatenate((img, np.array([c, c, c, 0], dtype=np.uint8)))
+            c = math.floor(((x+y)/(tree.ix+tree.iy))*255)
+            img = np.concatenate((img, np.array([c, c, c, 255], dtype=np.uint8)))
     await tree.update_image(img)
